Updated document preview from displaying pdf with iframe to react-pdf now displays in desktop and mobile view.
